Restructure auth home and basic components
