+++ guru-react/src/pages/Home/index.tsx
@@ -1,141 +1,119 @@
-import React, { Component } from "react";
-import { withRouter, RouteComponentProps } from "react-router-dom";
+import React, { useState } from "react";
+import { useHistory } from "react-router-dom";
+import { useSelector } from "react-redux";
 import axios from "axios";
-import { withStyles, WithStyles, createStyles } from "@material-ui/styles";
+
+import { makeStyles } from "@material-ui/styles";
 import { Theme } from "@material-ui/core/styles";
 import Grid from "@material-ui/core/Grid";
 import Typography from "@material-ui/core/Typography";
+
 import { colors } from "../../theme";
 import { paths } from "../../routes";
-import Page from "../../components/Page";
-import Navbar from "../../components/Navbar";
-import Footer from "../../components/Footer";
+import Page from "../../components/EmptyPage";
 import Autocomplete, {
   Suggestion,
   SearchButtonOptions
 } from "../../components/Autocomplete";
 import CategoryItem from "./CategoryItem";
-import { withInfo, WithInfoProps } from "../../state/hocs/info";
 
 import HeroImg from "../../assets/images/hero.svg";
 import { TicketSearchResult } from "../../state/types/tickets";
+import { AppState } from "../../state/types/state";
 
-const styles = (theme: Theme) =>
-  createStyles({
-    root: {},
-    searchHeader: {
-      backgroundImage: `url(${HeroImg})`,
-      backgroundSize: "cover",
-      backgroundRepeat: "no-repeat",
-      height: "23em",
-      marginBottom: "4em",
-      [theme.breakpoints.down("xs")]: {
-        height: "10em",
-        backgroundImage: "none"
-      }
-    },
-    searchInput: {
-      border: "none",
-      boxShadow: `5px 3px 4px 2px ${colors.VERY_LIGHT_TEXT}`
+const useStyles = makeStyles((theme: Theme) => ({
+  searchHeader: {
+    backgroundImage: `url(${HeroImg})`,
+    backgroundSize: "cover",
+    backgroundRepeat: "no-repeat",
+    height: "23em",
+    marginBottom: "4em",
+    [theme.breakpoints.down("xs")]: {
+      height: "10em",
+      backgroundImage: "none"
     }
-  });
-
-type Props = WithStyles<typeof styles> & RouteComponentProps & WithInfoProps;
+  },
+  searchInput: {
+    border: "none",
+    boxShadow: `5px 3px 4px 2px ${colors.VERY_LIGHT_TEXT}`
+  }
+}));
 
 type SearchSuggestion = TicketSearchResult & Suggestion;
 
-interface State {
-  searchResults: Array<SearchSuggestion>;
-  searchQuery: string;
-}
+const Home = () => {
+  const classes = useStyles();
+  const [searchResults, setSearchResults] = useState<Array<SearchSuggestion>>(
+    []
+  );
+  const history = useHistory();
+  const { categories } = useSelector((state: AppState) => state.info);
 
-class Home extends Component<Props, State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = {
-      searchResults: [],
-      searchQuery: ""
-    };
-  }
-
-  searchTickets = (q: string) => {
+  const searchTickets = (q: string) => {
     const url = `${process.env.REACT_APP_API_BASE}/api/v1/tickets/search/`;
     const params = { q };
 
     axios.get(url, { params }).then(response => {
-      this.setState({
-        searchResults: response.data.results
+      setSearchResults(
+        response.data.results
           .map((result: TicketSearchResult) => ({
             ...result,
             text: result.title
           }))
           .slice(0, 5)
-      });
+      );
     });
   };
 
-  handleSubmit = (selectedItem: Suggestion) => {
+  const handleSubmit = (selectedItem: Suggestion) => {
     const query = selectedItem.text;
-    this.props.history.push(`${paths.TICKET_LIST}?q=${query}`);
+    history.push(`${paths.TICKET_LIST}?q=${query}`);
   };
 
-  render() {
-    const { classes, info } = this.props;
-    const { searchResults } = this.state;
-    const { categories } = info;
+  const InputProps = {
+    classes: { notchedOutline: classes.searchInput }
+  };
 
-    const InputProps = {
-      classes: { notchedOutline: classes.searchInput }
-    };
-
-    return (
-      <Page>
-        <Navbar />
-        <div className={`${classes.root} app-body`}>
-          <div>
-            <Grid
-              container
-              justify="center"
-              alignItems="center"
-              classes={{ container: classes.searchHeader }}
-            >
-              <Grid item xs={10} md={6}>
-                <Typography variant="h5" align="center">
-                  Search for a Ticket
-                  <Autocomplete
-                    suggestions={searchResults}
-                    InputProps={InputProps}
-                    selectFunction={this.handleSubmit}
-                    updateQueryFunction={this.searchTickets}
-                    searchButton={SearchButtonOptions.Start}
-                    isAbsolute={true}
-                  />
-                </Typography>
-              </Grid>
+  return (
+    <Page>
+      <Grid
+        container
+        justify="center"
+        alignItems="center"
+        classes={{ container: classes.searchHeader }}
+      >
+        <Grid item xs={10} md={6}>
+          <Typography variant="h5" align="center">
+            Search for a Ticket
+            <Autocomplete
+              suggestions={searchResults}
+              InputProps={InputProps}
+              selectFunction={handleSubmit}
+              updateQueryFunction={searchTickets}
+              searchButton={SearchButtonOptions.Start}
+              isAbsolute={true}
+            />
+          </Typography>
+        </Grid>
+      </Grid>
+      <Grid container spacing={1} justify="center">
+        <Grid item xs={12} md={8}>
+          <Grid container spacing={3}>
+            <Grid item xs={12}>
+              <Typography variant="h5" align="center">
+                Select the category to find the related topic
+              </Typography>
             </Grid>
-          </div>
-          <Grid container spacing={1} justify="center">
-            <Grid item xs={12} md={8}>
-              <Grid container spacing={3}>
-                <Grid item xs={12}>
-                  <Typography variant="h5" align="center">
-                    Select the category to find the related topic
-                  </Typography>
-                </Grid>
-                {categories.map(category => (
-                  <Grid item key={category.id} xs={6} md={3}>
-                    <CategoryItem category={category} />
-                  </Grid>
-                ))}
+            {categories.map(category => (
+              <Grid item key={category.id} xs={6} md={3}>
+                <CategoryItem category={category} />
               </Grid>
-            </Grid>
+            ))}
           </Grid>
-        </div>
+        </Grid>
+      </Grid>
+    </Page>
+  );
+};
 
-        <Footer />
-      </Page>
-    );
-  }
-}
-
-export default withInfo(withStyles(styles)(withRouter(Home)));
+export default Home;
