+++ guru-react/src/pages/TicketDetail/Responses/Post/Responsive/index.tsx
@@ -8,9 +8,6 @@ import Button from "@material-ui/core/Button";
 import MenuItem from "@material-ui/core/MenuItem";
 import TextField from "@material-ui/core/TextField";
 import Grid from "@material-ui/core/Grid";
-import Box from "@material-ui/core/Box";
-import Card from "@material-ui/core/Card";
-import CardContent from "@material-ui/core/CardContent";
 
 import { colors } from "../../../../../theme";
 import useFunctions from "../hooks";
