+++ guru-react/src/pages/TicketDetail/TicketCard/Responsive/index.tsx
