+++ guru-react/src/pages/TicketDetail/TicketCard/index.tsx
