+++ guru-react/src/pages/TicketDetail/UserInfo/Responsive/index.tsx
